+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ClientInfo.kt
@@ -1,9 +1,10 @@
 @file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
+package com.sourcegraph.cody.protocol_generated;
 
 data class ClientInfo(
   val name: String,
   val version: String,
+  val ideVersion: String? = null,
   val workspaceRootUri: String,
   val workspaceRootPath: String? = null,
   val extensionConfiguration: ExtensionConfiguration? = null,