+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ContextItem.kt
@@ -1,18 +1,18 @@
 @file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
+package com.sourcegraph.cody.protocol_generated;
 
-import com.google.gson.annotations.SerializedName
-import com.google.gson.Gson
-import com.google.gson.JsonDeserializationContext
-import com.google.gson.JsonDeserializer
-import com.google.gson.JsonElement
-import java.lang.reflect.Type
+import com.google.gson.annotations.SerializedName;
+import com.google.gson.Gson;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import java.lang.reflect.Type;
 
 sealed class ContextItem {
   companion object {
     val deserializer: JsonDeserializer<ContextItem> =
       JsonDeserializer { element: JsonElement, _: Type, context: JsonDeserializationContext ->
-        when (element.asJsonObject.get("type").asString) {
+        when (element.getAsJsonObject().get("type").getAsString()) {
           "file" -> context.deserialize<ContextItemFile>(element, ContextItemFile::class.java)
           "repository" -> context.deserialize<ContextItemRepository>(element, ContextItemRepository::class.java)
           "tree" -> context.deserialize<ContextItemTree>(element, ContextItemTree::class.java)