+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/RemoteRepoFetchState.kt
@@ -1,7 +1,7 @@
 @file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
+package com.sourcegraph.cody.protocol_generated;
 
-import com.google.gson.annotations.SerializedName
+import com.google.gson.annotations.SerializedName;
 
 data class RemoteRepoFetchState(
   val state: StateEnum, // Oneof: paused, fetching, errored, complete