+++ vscode/src/completions/request-manager.test.ts
@@ -4,6 +4,7 @@ import { nextTick } from '@sourcegraph/cody-shared'
 import { getCurrentDocContext } from './get-current-doc-context'
 import { InlineCompletionsResultSource, TriggerKind } from './get-inline-completions'
 import { initCompletionProviderConfig } from './get-inline-completions-tests/helpers'
+import type { CompletionLogID } from './logger'
 import type { FetchCompletionResult } from './providers/fetch-and-process-completions'
 import { STOP_REASON_HOT_STREAK } from './providers/hot-streak'
 import { Provider } from './providers/provider'
@@ -76,6 +77,7 @@ function createProvider(prefix: string) {
         n: 1,
         firstCompletionTimeout: 1500,
         triggerKind: TriggerKind.Automatic,
+        completionLogId: 'mock-log-id' as CompletionLogID,
     })
 }