+++ vscode/test/e2e/chat-atFile.test.ts
@@ -1,5 +1,5 @@
-import { expect } from '@playwright/test'
 import { isWindows } from '@sourcegraph/cody-shared'
+import { expect } from 'playwright/test'
 import {
     atMentionMenuMessage,
     chatInputMentions,
@@ -101,13 +101,13 @@ test.extend<ExpectedEvents>({
     // Searching and clicking
     await chatInput.fill('Explain @mj')
     await chatPanelFrame.getByRole('option', { name: 'Main.java' }).click()
-    await expect(chatInput).toHaveText('Explain @Main.java ')
+    await expect(chatInput).toHaveText('Explain Main.java ')
     await chatInput.press('Enter')
     await expect(chatInput).toBeEmpty()
-    await expect(chatPanelFrame.getByText('Explain @Main.java')).toBeVisible()
+    await expect(chatPanelFrame.getByText('Explain Main.java')).toBeVisible()
     const contextCell = getContextCell(chatPanelFrame)
     await expect(contextCell).toHaveCount(1)
-    await expect(chatInput).not.toHaveText('Explain @Main.java ')
+    await expect(chatInput).not.toHaveText('Explain Main.java ')
     await expect(chatPanelFrame.getByRole('option', { name: 'Main.java' })).not.toBeVisible()
 
     // Keyboard nav through context files
@@ -117,7 +117,7 @@ test.extend<ExpectedEvents>({
     ).toBeVisible()
     await chatInput.press('Tab')
     await chatInput.press('Tab')
-    await expect(chatInput).toHaveText(withPlatformSlashes('Explain @var.go '))
+    await expect(chatInput).toHaveText(withPlatformSlashes('Explain var.go '))
     await chatInput.focus()
     await chatInput.pressSequentially('and ')
     await chatInput.pressSequentially('@vgo', { delay: 10 })
@@ -130,13 +130,14 @@ test.extend<ExpectedEvents>({
     await chatInput.press('ArrowDown') // second item again
     await expect(chatPanelFrame.getByRole('option', { selected: true })).toHaveText(/visualize\.go/)
     await chatInput.press('Tab')
-    await expect(chatInput).toHaveText(withPlatformSlashes('Explain @var.go and @visualize.go '))
+    await expect(chatInput).toHaveText(withPlatformSlashes('Explain var.go and visualize.go '))
+    await expect(chatInputMentions(chatInput)).toHaveText(['var.go', 'visualize.go'])
 
     // Send the message and check it was included
     await chatInput.press('Enter')
     await expect(chatInput).toBeEmpty()
     await expect(
-        chatPanelFrame.getByText(withPlatformSlashes('Explain @var.go and @visualize.go'))
+        chatPanelFrame.getByText(withPlatformSlashes('Explain var.go and visualize.go'))
     ).toBeVisible()
 
     // Ensure explicitly @-included context shows up as enhanced context
@@ -149,14 +150,16 @@ test.extend<ExpectedEvents>({
     await chatInput.pressSequentially('@Main.java', { delay: 10 })
     await expect(chatPanelFrame.getByRole('option', { name: 'Main.java' })).toBeVisible()
     await chatInput.press('Tab')
-    await expect(chatInput).toHaveText('@Main.java ')
+    await expect(chatInput).toHaveText('Main.java ')
+    await expect(chatInputMentions(chatInput)).toHaveText(['Main.java'])
 
     // Check pressing tab after typing a partial filename but where that complete
     // filename already exists earlier in the input.
     // https://github.com/sourcegraph/cody/issues/2243
     await chatInput.pressSequentially('and @Main.ja', { delay: 10 })
     await chatInput.press('Tab')
-    await expect(chatInput).toHaveText('@Main.java and @Main.java ')
+    await expect(chatInput).toHaveText('Main.java and Main.java ')
+    await expect(chatInputMentions(chatInput)).toHaveText(['Main.java', 'Main.java'])
 
     // Support @-file in mid-sentence
     await chatInput.focus()
@@ -171,16 +174,16 @@ test.extend<ExpectedEvents>({
     await chatInput.pressSequentially('@Main', { delay: 10 })
     await expect(chatPanelFrame.getByRole('option', { name: 'Main.java' })).toBeVisible()
     await chatInput.press('Tab')
-    await expect(chatInput).toHaveText('Explain the @Main.java file')
+    await expect(chatInput).toHaveText('Explain the Main.java file')
     // Confirm the cursor is at the end of the newly added file name with space
     await page.keyboard.press('!')
     await page.keyboard.press('Delete')
-    await expect(chatInput).toHaveText('Explain the @Main.java !file')
+    await expect(chatInput).toHaveText('Explain the Main.java !file')
 
     //  "ArrowLeft" / "ArrowRight" keys alter the query input for @-mentions.
     const noMatches = atMentionMenuMessage(chatPanelFrame, 'No files found')
     await chatInput.pressSequentially(' @abcdefg')
-    await expect(chatInput).toHaveText('Explain the @Main.java ! @abcdefgfile')
+    await expect(chatInput).toHaveText('Explain the Main.java ! @abcdefgfile')
     await noMatches.hover()
     await expect(noMatches).toBeVisible()
     await chatInput.press('ArrowLeft')
@@ -188,7 +191,7 @@ test.extend<ExpectedEvents>({
     await chatInput.press('ArrowRight')
     await expect(noMatches).toBeVisible()
     await chatInput.press('$')
-    await expect(chatInput).toHaveText('Explain the @Main.java ! @abcdefg$file')
+    await expect(chatInput).toHaveText('Explain the Main.java ! @abcdefg$file')
     await expect(noMatches).not.toBeVisible()
     // Selection close on submit
     await chatInput.press('Enter')
@@ -247,24 +250,25 @@ test.extend<ExpectedEvents>({
     // Send a message with an @-mention.
     await firstChatInput.fill('Explain @mj')
     await chatPanelFrame.getByRole('option', { name: 'Main.java' }).click()
-    await expect(firstChatInput).toHaveText('Explain @Main.java ')
+    await expect(firstChatInput).toHaveText('Explain Main.java ')
     await firstChatInput.press('Enter')
     const contextCell = getContextCell(chatPanelFrame)
     await expectContextCellCounts(contextCell, { files: 1 })
 
     // Edit the just-sent message and resend it. Confirm it is sent with the right context items.
-    await expect(firstChatInput).toHaveText('Explain @Main.java ')
+    await expect(firstChatInput).toHaveText('Explain Main.java ')
     await firstChatInput.press('Meta+Enter')
     await expectContextCellCounts(contextCell, { files: 1 })
 
     // Edit it again, add a new @-mention, and resend.
-    await expect(firstChatInput).toHaveText('Explain @Main.java ')
+    await expect(firstChatInput).toHaveText('Explain Main.java ')
     await focusChatInputAtEnd(firstChatInput)
     await firstChatInput.pressSequentially('and @index.ht')
     await chatPanelFrame.getByRole('option', { name: 'index.html' }).click()
-    await expect(firstChatInput).toHaveText('Explain @Main.java and @index.html')
+    await expect(firstChatInput).toHaveText('Explain Main.java and index.html')
     await firstChatInput.press('Enter')
-    await expect(firstChatInput).toHaveText('Explain @Main.java and @index.html')
+    await expect(firstChatInput).toHaveText('Explain Main.java and index.html')
+    await expect(chatInputMentions(firstChatInput)).toHaveText(['Main.java', 'index.html'])
     await expectContextCellCounts(contextCell, { files: 2 })
 })
 
@@ -301,7 +305,7 @@ test.extend<ExpectedEvents>({
     await chatInput.fill('@buzz.ts:2-4')
     await expect(chatPanelFrame.getByRole('option', { name: 'buzz.ts Lines 2-4' })).toBeVisible()
     await chatPanelFrame.getByRole('option', { name: 'buzz.ts Lines 2-4' }).click()
-    await expect(chatInput).toHaveText('@buzz.ts:2-4 ')
+    await expect(chatInput).toHaveText('buzz.ts:2-4 ')
     // Submit the message
     await chatInput.press('Enter')
 
@@ -377,7 +381,8 @@ test.extend<ExpectedEvents>({
     await chatInput.pressSequentially('fizzb', { delay: 10 })
     await expect(chatPanelFrame.getByRole('option', { name: 'fizzbuzz()' })).toBeVisible()
     await chatPanelFrame.getByRole('option', { name: 'fizzbuzz()' }).click()
-    await expect(chatInput).toHaveText('@buzz.ts @fizzbuzz()')
+    await expect(chatInput).toHaveText('buzz.ts fizzbuzz() ')
+    await expect(chatInputMentions(chatInput)).toHaveText(['buzz.ts', 'fizzbuzz()'])
 
     // Submit the message
     await chatInput.press('Enter')
@@ -413,5 +418,5 @@ test.extend<ExpectedEvents>({
     await lastChatInput.press('x')
     await selectLineRangeInEditorTab(page, 7, 10)
     await executeCommandInPalette(page, 'Cody: Add Selection to Cody Chat')
-    await expect(chatInputMentions(lastChatInput)).toHaveText(['@buzz.ts:2-5', '@buzz.ts:7-10'])
+    await expect(chatInputMentions(lastChatInput)).toHaveText(['buzz.ts:2-5', 'buzz.ts:7-10'])
 })