+++ vscode/webviews/mentions/mentionMenu/MentionMenuItem.tsx
@@ -116,7 +116,7 @@ export const MentionMenuProviderItemContent: FunctionComponent<{
     )
 }
 
-const iconForProvider: Record<
+export const iconForProvider: Record<
     string,
     React.ComponentType<{
         size?: string | number